fix to register real
i forgot i cant use middleware anymore whoopsies
